CapGen/requirements.txt
gradio 
openai 
pillow 
torchvision 
dotenv
transformers
